+++ test/helper/init.lua
@@ -1,0 +1,7 @@
+local utils = require('test.helper.utils')
+
+if utils.is_tarantool_3() then
+    return require('test.helper.config_tarantool')
+else
+    return require('test.helper.config_cartridge')
+end